🧹 Cleanup 🔨 Remove 'lazy_static'
