+++ macchina/Cargo.toml
@@ -17,6 +17,5 @@ tui = { version = "0.14.0", default-features = false, features = ['crossterm'] }
 colored = "2.0.0"
 structopt = "0.3.21"
 rand = "0.8.3"
-lazy_static = "1.4.0"
 unicode-width = "0.1.8"
 macchina-read = { version = "0.1.3", path = "../macchina-read" }
